requirements.txt
absl-py==2.0.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.11
cached-property==1.5.2
cachetools==5.3.2
catalogue==2.0.10
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
comm==0.1.4
confection==0.1.3
cycler==0.11.0
cymem==2.0.8
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl
debugpy==1.7.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
entrypoints==0.4
exceptiongroup==1.2.0
fastjsonschema==2.19.0
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
keras==2.11.0
Keras-Applications==1.0.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.0.1
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
murmurhash==1.0.10
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.8.0
nest-asyncio==1.5.8
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==23.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.41
protobuf==3.19.6
psutil==5.9.6
ptyprocess==0.7.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
Pygments==2.17.1
pyparsing==3.1.1
pyrsistent==0.19.3
python-dateutil==2.8.2
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.7.3
Send2Trash==1.8.2
six==1.15.0
smart-open==6.4.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
swin-window-process==0.0.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-addons==0.19.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==1.1.0
terminado==0.17.1
thinc==8.2.1
timm==0.4.12
tinycss2==1.2.1
torch==1.8.0
torchvision==0.9.0
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typeguard==4.1.2
typer==0.9.0
typing_extensions==4.7.1
urllib3==2.0.7
wasabi==1.1.2
wcwidth==0.2.11
weasel==0.3.4
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.9
wrapt==1.12.1
yacs==0.1.8
zipp==3.15.0
